mv `export` to new line
having `export` on the same line as a variable declaration that uses `$()` can mask any exit codes that occur in the `$()` statement since `export` will usually still exit with `0`.

this issue was causing this line to fail silently when it was trying to retrieve branch SHAs for private repos, but didn't have permission to do so with the default `GH_TOKEN`.

see the shellcheck page below for more details.

- https://www.shellcheck.net/wiki/SC2155