Make tests run when builds pass (#15)
Currently, tests get skipped if a transitive dependency fails; it is not
clear why this is the case, given the applied condition of `if: ${{
needs.<job-id>.result == 'success' }}` for the test jobs.

For instance, `cudf-tests` depend on `cudf-build`, which depends on
`rmm-build` -- now if `rmm-build` fails, even if `cudf-build` passes,
`cudf-tests` inevitably get skipped. See [this
example](https://github.com/rapidsai/actions/actions/runs/4868448744).

This is not the desired workflow. The desired workflow is to have the
test jobs run as long as their immediate build dependencies pass. This
PR ensures this.

[This
workflow](https://github.com/AyodeAwe/testing/actions/runs/4888641367)
demonstrates this change.