enable cugraph-docs nightly builds (#68)
Starting with the 25.02 release, docs building for all of the cuGraph
projects (e.g. `cugraph`, `nx-cugraph`, `wholegraph`, etc.) has moved to
a single repo: https://github.com/rapidsai/cugraph-docs.

Now that https://github.com/rapidsai/cugraph/pull/4837 was merged, the
`cugraph-docs` repo is the sole source of docs for those projects.

This proposes triggering its builds as part of the nightly pipeline, to
keep the docs site up to date with changes to the cuGraph projects.